CRUD for subscriptions done!

+++ src/main/java/com/diplom/backend/user/entity/User.java
@@ -1,18 +1,20 @@
 package com.diplom.backend.user.entity;
 
 import com.diplom.backend.subscription.entity.Subscription;
+import org.bson.types.ObjectId;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.index.Indexed;
 import org.springframework.data.mongodb.core.mapping.DBRef;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Document(collection = "user")
 public class User {
 
     @Id
-    private long id;
+    private BigInteger id;
 
     @Indexed(unique = true)
     private String username;
@@ -44,11 +46,11 @@ public class User {
         this.password = password;
     }
 
-    public long getId() {
+    public BigInteger getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(BigInteger id) {
         this.id = id;
     }
